#400 - [REFACTOR] Infrastructure 영역의 디렉터리 구조가 기술 중심의 구조여서 도메인 중심의 구조로 변경했습니다.
