+++ apps/server-stock/src/main/java/com/assetmind/server_stock/global/config/JpaConfig.java
@@ -4,7 +4,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 @Configuration
-@EnableJpaRepositories(basePackages = "com.assetmind.server_stock.stock.infrastructure.persistence.jpa")
+@EnableJpaRepositories(basePackages = "com.assetmind.server_stock.stock.infrastructure.persistence")
 public class JpaConfig {
 
 }
