+++ apps/server-stock/src/main/java/com/assetmind/server_stock/global/config/RedisConfig.java
@@ -4,7 +4,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.data.redis.repository.configuration.EnableRedisRepositories;
 
 @Configuration
-@EnableRedisRepositories(basePackages = "com.assetmind.server_stock.stock.infrastructure.persistence.redis")
+@EnableRedisRepositories(basePackages = "com.assetmind.server_stock.stock.infrastructure.persistence")
 public class RedisConfig {
 
 }
