+++ apps/server-stock/src/main/java/com/assetmind/server_stock/stock/infrastructure/persistence/common/keys/TickId.java
@@ -1,4 +1,4 @@
-package com.assetmind.server_stock.stock.infrastructure.persistence.entity.keys;
+package com.assetmind.server_stock.stock.infrastructure.persistence.common.keys;
 
 import java.io.Serializable;
 import java.time.LocalDateTime;
