+++ apps/server-stock/src/main/java/com/assetmind/server_stock/stock/infrastructure/persistence/ohlcv/adapter/Ohlcv1dJpaAdapter.java
@@ -1,9 +1,10 @@
-package com.assetmind.server_stock.stock.infrastructure.persistence.jpa;
+package com.assetmind.server_stock.stock.infrastructure.persistence.ohlcv.adapter;
 
 import com.assetmind.server_stock.stock.domain.dtos.OhlcvDto;
 import com.assetmind.server_stock.stock.domain.repository.Ohlcv1dRepository;
-import com.assetmind.server_stock.stock.infrastructure.persistence.entity.Ohlcv1dJpaEntity;
-import com.assetmind.server_stock.stock.infrastructure.persistence.entity.projection.ChartCandleProjection;
+import com.assetmind.server_stock.stock.infrastructure.persistence.ohlcv.entity.Ohlcv1dJpaEntity;
+import com.assetmind.server_stock.stock.infrastructure.persistence.common.projection.ChartCandleProjection;
+import com.assetmind.server_stock.stock.infrastructure.persistence.ohlcv.repository.Ohlcv1dJpaRepository;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
@@ -15,7 +16,7 @@ import org.springframework.stereotype.Repository;
  * {@link Ohlcv1dRepository} 인터페이스를 JPA 기술로 구현한 어탭터 구현체
  *
  * 도메인 계층에서 전달받은 순수 DTO {@link OhlcvDto}를
- * 영속성 객체 {@link com.assetmind.server_stock.stock.infrastructure.persistence.entity.Ohlcv1dJpaEntity}로 변환하여 DB에 저장
+ * 영속성 객체 {@link Ohlcv1dJpaEntity}로 변환하여 DB에 저장
  *
  */
 @Repository
