+++ apps/server-stock/src/main/java/com/assetmind/server_stock/stock/infrastructure/persistence/ohlcv/entity/Ohlcv1dJpaEntity.java
@@ -1,6 +1,6 @@
-package com.assetmind.server_stock.stock.infrastructure.persistence.entity;
+package com.assetmind.server_stock.stock.infrastructure.persistence.ohlcv.entity;
 
-import com.assetmind.server_stock.stock.infrastructure.persistence.entity.keys.OhlcvId;
+import com.assetmind.server_stock.stock.infrastructure.persistence.common.keys.OhlcvId;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
@@ -14,10 +14,10 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @Entity
-@Table(name = "ohlcv_1m")
+@Table(name = "ohlcv_1d")
 @IdClass(OhlcvId.class)
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Ohlcv1mJpaEntity {
+public class Ohlcv1dJpaEntity {
 
     @Id
     @Column(name = "stock_code", nullable = false, length = 20)
@@ -43,7 +43,7 @@ public class Ohlcv1mJpaEntity {
     private Long volume;
 
     @Builder
-    public Ohlcv1mJpaEntity(String stockCode, LocalDateTime candleTimestamp,
+    public Ohlcv1dJpaEntity(String stockCode, LocalDateTime candleTimestamp,
             Double openPrice, Double highPrice, Double lowPrice, Double closePrice, Long volume) {
         this.stockCode = stockCode;
         this.candleTimestamp = candleTimestamp;
@@ -53,5 +53,4 @@ public class Ohlcv1mJpaEntity {
         this.closePrice = closePrice;
         this.volume = volume;
     }
-
 }
