+++ apps/server-stock/src/main/java/com/assetmind/server_stock/stock/infrastructure/persistence/stock/entity/StockMetaEntity.java
@@ -1,4 +1,4 @@
-package com.assetmind.server_stock.stock.infrastructure.persistence.entity;
+package com.assetmind.server_stock.stock.infrastructure.persistence.stock.entity;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
