+++ apps/server-stock/src/main/java/com/assetmind/server_stock/stock/infrastructure/persistence/stock/repository/StockPriceRedisRepository.java
@@ -1,6 +1,6 @@
-package com.assetmind.server_stock.stock.infrastructure.persistence.redis;
+package com.assetmind.server_stock.stock.infrastructure.persistence.stock.repository;
 
-import com.assetmind.server_stock.stock.infrastructure.persistence.entity.StockPriceRedisEntity;
+import com.assetmind.server_stock.stock.infrastructure.persistence.stock.entity.StockPriceRedisEntity;
 import org.springframework.data.repository.CrudRepository;
 
 /**
