+++ apps/server-stock/src/main/java/com/assetmind/server_stock/stock/presentation/dto/StockHistoryResponse.java
@@ -1,7 +1,7 @@
 package com.assetmind.server_stock.stock.presentation.dto;
 
 import com.assetmind.server_stock.stock.application.listener.dto.RealTimeStockTradeEvent;
-import com.assetmind.server_stock.stock.infrastructure.persistence.entity.RawTickJpaEntity;
+import com.assetmind.server_stock.stock.infrastructure.persistence.tick.entity.RawTickJpaEntity;
 import java.time.format.DateTimeFormatter;
 import lombok.Builder;
