+++ apps/server-stock/src/test/java/com/assetmind/server_stock/stock/application/StockServiceTest.java
@@ -19,8 +19,8 @@ import com.assetmind.server_stock.stock.domain.repository.CandleRepository;
 import com.assetmind.server_stock.stock.domain.repository.RawTickRepository;
 import com.assetmind.server_stock.stock.domain.repository.StockSnapshotRepository;
 import com.assetmind.server_stock.stock.exception.StockNotFoundException;
-import com.assetmind.server_stock.stock.infrastructure.persistence.entity.RawTickJpaEntity;
-import com.assetmind.server_stock.stock.infrastructure.persistence.entity.StockPriceRedisEntity;
+import com.assetmind.server_stock.stock.infrastructure.persistence.tick.entity.RawTickJpaEntity;
+import com.assetmind.server_stock.stock.infrastructure.persistence.stock.entity.StockPriceRedisEntity;
 import com.assetmind.server_stock.stock.presentation.dto.StockHistoryResponse;
 import com.assetmind.server_stock.stock.presentation.dto.StockRankingResponse;
 import java.time.LocalDateTime;
