+++ apps/server-stock/src/test/java/com/assetmind/server_stock/stock/infrastructure/persistence/jdbc/PartitionJdbcAdapterTest.java
@@ -7,6 +7,7 @@ import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
+import com.assetmind.server_stock.stock.infrastructure.persistence.tick.adapter.PartitionJdbcAdapter;
 import java.time.LocalDate;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
