+++ apps/server-stock/src/test/java/com/assetmind/server_stock/stock/infrastructure/persistence/redis/RedisAlertThrottlingAdapterTest.java
@@ -5,6 +5,7 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.when;
 
+import com.assetmind.server_stock.stock.infrastructure.persistence.alert.RedisAlertThrottlingAdapter;
 import java.time.Duration;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
