+++ apps/server-stock/src/test/java/com/assetmind/server_stock/stock/integration/StockRestApiIntegrationTest.java
@@ -11,8 +11,8 @@ import static org.springframework.restdocs.request.RequestDocumentation.*;
 import com.assetmind.server_stock.stock.application.provider.StockMetadataProvider;
 import com.assetmind.server_stock.stock.domain.repository.RawTickRepository;
 import com.assetmind.server_stock.stock.domain.repository.StockSnapshotRepository;
-import com.assetmind.server_stock.stock.infrastructure.persistence.entity.RawTickJpaEntity;
-import com.assetmind.server_stock.stock.infrastructure.persistence.entity.StockPriceRedisEntity;
+import com.assetmind.server_stock.stock.infrastructure.persistence.tick.entity.RawTickJpaEntity;
+import com.assetmind.server_stock.stock.infrastructure.persistence.stock.entity.StockPriceRedisEntity;
 import com.assetmind.server_stock.support.IntegrationTestSupport;
 import java.time.LocalDateTime;
 import org.junit.jupiter.api.AfterEach;
